move testdata; add gitignore; adjust Readme
