+++ representer/testdata/hamming/1/embed.go
