+++ representer/testdata/hamming/1/mapping.json
