+++ representer/testdata/hamming/1/solution.go
