+++ representer/testdata/hamming/2/mapping.json
