+++ representer/testdata/hamming/2/solution.go
