+++ representer/testdata/hamming/3/mapping.json
