+++ representer/testdata/hamming/3/representation.txt
