+++ representer/testdata/hamming/3/solution.go
