+++ representer/testdata/hamming/4/embed.go
