+++ representer/testdata/hamming/4/mapping.json
