+++ representer/testdata/hamming/4/representation.txt
