+++ representer/testdata/hamming/4/solution.go
