+++ representer/testdata/raindrops/3/solution.go
